Interface Created
A basic terminal based interface has been implemented. The user can choose to use the program in encode or decode mode (or just exit).
These modes allow the user to input a word or code and it will be encoded or decoded respectively.
Still needs some sort of input validation.